Rewriting driver class
Minor changes to quadrilaterals

+++ src/Quad/Kite.java
@@ -1,12 +1,12 @@
 package Quad;
 
 public class Kite extends Quadrilateral{
-    private int side1;
-    private int side2;
-    private int diag1;
-    private int diag2;
+    private double side1;
+    private double side2;
+    private double diag1;
+    private double diag2;
     private String key;
-    private static int kiteNum;
+    private static double kiteNum;
     public Kite(){
         super();
         side1 = 0;
@@ -18,7 +18,7 @@ public class Kite extends Quadrilateral{
             key = "Kite " + kiteNum;
         }
     }
-    public Kite(int side1, int side2, int diag1, int diag2){
+    public Kite(double side1, double side2, double diag1, double diag2){
         this.side1 = side1;
         this.side2 = side2;
         this.diag1 = diag1;
@@ -29,26 +29,26 @@ public class Kite extends Quadrilateral{
         }
     }
 
-    public void setSide1(int side1) { this.side1 = side1; }
-    public void setSide2(int side2) { this.side2 = side2; }
-    public void setDiag1(int diag1) { this.diag1 = diag1; }
-    public void setDiag2(int diag2) { this.diag2 = diag2; }
+    public void setSide1(double side1) { this.side1 = side1; }
+    public void setSide2(double side2) { this.side2 = side2; }
+    public void setDiag1(double diag1) { this.diag1 = diag1; }
+    public void setDiag2(double diag2) { this.diag2 = diag2; }
 
-    public int getSide1() {
+    public double getSide1() {
         return side1;
     }
-    public int getSide2() {
+    public double getSide2() {
         return side2;
     }
-    public int getDiag1() {
+    public double getDiag1() {
         return diag1;
     }
-    public int getDiag2() {
+    public double getDiag2() {
         return diag2;
     }
     public String getKey() { return key; }
 
-    public int getNum() { return kiteNum; }
+    public double getNum() { return kiteNum; }
 
     public void remove() {
         if (getClass() == Kite.class) {
@@ -62,11 +62,11 @@ public class Kite extends Quadrilateral{
         return "<Kite> Side 1:" + getSide1() + ", Side 2: " + getSide2() + ", Diagonal 1: " +  getDiag1() + ", Diagonal 2: " + getDiag2();
     }
     @Override
-    public int findArea(){
+    public double findArea(){
         return getDiag1() * getDiag2() / 2;
     }
     @Override
-    public int findParameter() {
+    public double findParameter() {
         return (getSide1() + getSide2()) * 2;
     }
     @Override
