+++ src/Quad/Quadrilateral.java
@@ -1,7 +1,7 @@
 package Quad;
 
 public abstract class Quadrilateral {
-    private static int quadNum;
+    private static double quadNum;
     public Quadrilateral(){
         quadNum += 1;
     }
@@ -14,11 +14,11 @@ public abstract class Quadrilateral {
      * But I think this way show better understanding of OOP
      */
     public void subtractAmount() { quadNum -= 1;}
-    public int getQuadNum() { return quadNum; }
+    public double getQuadNum() { return quadNum; }
 
     abstract void remove(); // When a shape is removed from data base
-    abstract int getNum();
-    abstract int findArea();
-    abstract int findParameter();
+    abstract double getNum();
+    abstract double findArea();
+    abstract double findParameter();
     abstract String getKey();
 }
